requirements.txt
alabaster==0.7.12
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
CommonMark==0.5.4
cryptography==1.7.1
docutils==0.14
Flask==1.0.2
idna==2.8
imagesize==1.1.0
isc-dhcp-leases==0.9.1
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.1
keyrings.alt==1.3
MarkupSafe==1.1.0
mysql-connector==2.1.6
packaging==18.0
purepng==0.2.0
pyasn1==0.1.9
pycrypto==2.6.1
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.22.0
pyparsing==2.3.0
python-apt==1.1.0b5
pytz==2018.7
pyxdg==0.25
recommonmark==0.4.0
requests==2.21.0
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.3.1
SecretStorage==2.3.1
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
unattended-upgrades==0.1
urllib3==1.24.1
uWSGI==2.0.17.1
virtualenv==15.1.0
Werkzeug==0.14.1
